Choose more impactful code split points for improved user experience when navigating app
